// Frontend v2 (2)/Frontend v2/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        JobResponse: {
            /** Title */
            title: string;
            /** Company */
            company: string;
            /** Location */
            location: string;
            /** Type */
            type: string;
            /** Description */
            description: string;
            /** Requirements */
            requirements?: string[] | null;
            /** Salary */
            salary?: string | null;
            /** Poster Id */
            poster_id?: string | null;
            /**
             * Is Active
             * @default true
             */
            is_active: boolean;
            /** Application Deadline */
            application_deadline?: string | null;
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Created At */
            created_at: string;
            /** Updated At */
            updated_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}